+++ dio/lib/diox.dart
@@ -17,3 +17,5 @@ export 'src/parameter.dart';
 export 'src/redirect_record.dart';
 export 'src/response.dart';
 export 'src/transformer.dart';
+export 'src/transformers/background_transformer.dart';
+export 'src/transformers/sync_transformer.dart';